remove tasks ot required by raw provider, fix provider name in plugins after renaming from one to raw
