+++ providers/raw/tasks/boot.py
@@ -26,13 +26,8 @@ class ConfigureGrub(Task):
 
 		from common.tools import sed_i
 		grub_def = os.path.join(info.root, 'etc/default/grub')
-		#sed_i(grub_def, '^GRUB_TIMEOUT=[0-9]+', 'GRUB_TIMEOUT=0\n'
-		#                                        'GRUB_HIDDEN_TIMEOUT=true')
 
 		from common.tools import log_check_call
-		#log_check_call(['/usr/sbin/chroot', info.root, '/usr/sbin/update-grub'])
-		#log_check_call(['/usr/sbin/chroot', info.root, 'cat', '/boot/grub/grub.cfg'])
-		#log_check_call(['/usr/sbin/chroot', info.root, 'cat', '/boot/grub/device.map'])
 		log_check_call(['/usr/sbin/chroot', info.root, 'ln', '-s', '/boot/grub/grub.cfg', '/boot/grub/menu.lst'])
 
 		log_check_call(['/usr/sbin/chroot', info.root, 'update-initramfs', '-u'])
